feat: 🎸 separating routes
separating routes into custom files

✅ Closes: single responsability
