+++ routes/index.js
@@ -1,0 +1,8 @@
+const productsRouter = require('./products.router');
+
+function routerApi(app) {
+  app.use('/products', productsRouter);
+  // app.use('/users', usersRouter)
+}
+
+module.exports = routerApi;
